refactor: CommunitySwitcherをページから独立したコンポーネントに切り出し
