+++ src/app/community/[communityId]/users/me/page.tsx
@@ -7,26 +7,13 @@ import { useCommunityConfig } from "@/contexts/CommunityConfigContext";
 import { GqlMembership, GqlRole, useGetCommunitiesQuery } from "@/types/graphql";
 import useHeaderConfig from "@/hooks/useHeaderConfig";
 import { useMemo } from "react";
-import { AppLink } from "@/lib/navigation";
-import { Button } from "@/components/ui/button";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
-import { useTranslations } from "next-intl";
-import { ArrowLeftRight, Check } from "lucide-react";
-import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
+import { CommunitySwitcher } from "./CommunitySwitcher";
 
 export default function MyProfilePage() {
   const { gqlUser, isOwner, portfolios } = useUserProfileContext();
   const { user: currentUser } = useAuth();
   const communityConfig = useCommunityConfig();
   const communityId = communityConfig?.communityId;
-  const t = useTranslations();
 
   const memberships = useMemo(() => currentUser?.memberships ?? [], [currentUser?.memberships]);
   const joinedCommunityIds = useMemo(
@@ -55,58 +42,18 @@ export default function MyProfilePage() {
       .sort((a, b) => (a.name ?? "").localeCompare(b.name ?? ""));
   }, [communitiesData, joinedCommunityIds]);
 
-  // ヘッダー設定
   const headerConfig = useMemo(
     () => ({
       action: showSwitcher ? (
-        <DropdownMenu>
-          <DropdownMenuTrigger asChild>
-            <Button variant="tertiary" size="sm" disabled={communitiesLoading}>
-              {t("users.profileHeader.switchButton")}
-              <ArrowLeftRight className="w-4 h-4 ml-1" />
-            </Button>
-          </DropdownMenuTrigger>
-          <DropdownMenuContent align="end">
-            <DropdownMenuLabel>{t("users.profileHeader.switchLabel")}</DropdownMenuLabel>
-            <DropdownMenuSeparator />
-            {joinedCommunities.map((community) => {
-              const isCurrent = community.id === communityId;
-              return (
-                <DropdownMenuItem key={community.id} asChild disabled={isCurrent}>
-                  <AppLink href="/users/me" communityId={community.id} className="flex items-center gap-3 py-2">
-                    <div className="relative shrink-0">
-                      <Avatar className="h-8 w-8">
-                        <AvatarImage src={community.image ?? undefined} alt={community.name ?? ""} />
-                        <AvatarFallback className="text-xs font-medium">
-                          {community.name?.charAt(0) ?? "?"}
-                        </AvatarFallback>
-                      </Avatar>
-                      {isCurrent && (
-                        <span className="absolute -bottom-0.5 -right-0.5 flex h-4 w-4 items-center justify-center rounded-full bg-primary ring-2 ring-background">
-                          <Check className="h-2.5 w-2.5 text-primary-foreground" />
-                        </span>
-                      )}
-                    </div>
-                    <span className="flex-1 truncate text-sm">{community.name}</span>
-                  </AppLink>
-                </DropdownMenuItem>
-              );
-            })}
-            {hasAdminRole && (
-              <>
-                <DropdownMenuSeparator />
-                <DropdownMenuItem asChild>
-                  <AppLink href="/admin">
-                    {t("users.profileHeader.adminButton")}
-                  </AppLink>
-                </DropdownMenuItem>
-              </>
-            )}
-          </DropdownMenuContent>
-        </DropdownMenu>
+        <CommunitySwitcher
+          communities={joinedCommunities}
+          currentCommunityId={communityId}
+          hasAdminRole={hasAdminRole}
+          loading={communitiesLoading}
+        />
       ) : undefined,
     }),
-    [showSwitcher, joinedCommunities, communityId, hasAdminRole, t, communitiesLoading],
+    [showSwitcher, joinedCommunities, communityId, hasAdminRole, communitiesLoading],
   );
 
   useHeaderConfig(headerConfig);
